refactor(suite): one list wrapper and options context for trading pickers
Buy and sell each had a copy of the picker list shell, differing only in
memo and the item type, and a copy of the asset-options context, sell's
being buy's minus assets. Both now live in TradingFormInputAssetPicker,
the wrapper generic over the item type with a getItemHeight prop.

buildAssetOptions moves out of the provider into buy's
useBuildTradingAssetOptions, so the asset list is no longer built while
the picker is closed.
