+++ packages/suite/src/views/wallet/trading/common/TradingForm/TradingFormInput/TradingFormInputAssetPicker/AssetListWrapper/AssetListWrapper.tsx
@@ -1,26 +1,24 @@
 import { type ReactNode, memo, useRef } from 'react';
 
 import { AssetsList, AssetsListEmpty } from 'src/components/suite/asset-picker/components';
-import { ASSET_ROW_HEIGHT } from 'src/components/suite/asset-picker/constants';
 import { useListScrollReset } from 'src/components/suite/asset-picker/hooks';
-
-import { type TradingAssetListItem } from './hooks/useBuildTradingAssetOptions';
 
 const LIST_HEIGHT = 530;
 
-const getItemHeight = () => ASSET_ROW_HEIGHT;
-
-export interface AssetListWrapperProps {
-    renderItem: (item: TradingAssetListItem) => ReactNode;
-    listItems: TradingAssetListItem[];
+export interface AssetListWrapperProps<T> {
+    listItems: T[];
+    renderItem: (item: T) => ReactNode;
+    getItemHeight: (item: T) => number;
+    //Trigger to reset scroll position when this prop. changes
     resetScrollTrigger: string;
 }
 
-export const AssetListWrapper = memo(function AssetListWrapperInner({
+function AssetListWrapperInner<T>({
     listItems,
     renderItem,
+    getItemHeight,
     resetScrollTrigger,
-}: AssetListWrapperProps) {
+}: AssetListWrapperProps<T>) {
     const listRef = useRef<HTMLDivElement>(null);
 
     useListScrollReset(listRef, resetScrollTrigger);
@@ -41,4 +39,6 @@ export const AssetListWrapper = memo(function AssetListWrapperInner({
             />
         </AssetsListEmpty>
     );
-});
+}
+
+export const AssetListWrapper = memo(AssetListWrapperInner) as typeof AssetListWrapperInner;
