+++ packages/suite/src/views/wallet/trading/common/TradingForm/TradingFormInput/TradingFormInputAssetPicker/index.ts
@@ -1,3 +1,5 @@
 export * from './AssetPickerInput/AssetPickerInput';
 export * from './AssetPickerSearchHeader/AssetPickerSearchHeader';
 export * from './AssetPickerInput/AssetPickerInputBalance';
+export * from './AssetListWrapper/AssetListWrapper';
+export * from './AssetOptionsContext';
