--- packages/suite/src/views/wallet/trading/common/TradingForm/TradingFormInput/TradingFormInputBuyAsset/AssetOptionsContext.tsx
@@ -1,43 +1,0 @@
-import { type ReactNode, createContext, useContext, useMemo } from 'react';
-
-import { type CryptoId } from 'invity-api';
-
-import { type TradingAssetOption, useTradingAssets } from '@suite-common/trading';
-import { throwError } from '@trezor/utils';
-
-const AssetOptionsContext = createContext<{
-    assets: TradingAssetOption[];
-    includedCryptoIds: Set<CryptoId>;
-    excludedCryptoIds: Set<CryptoId>;
-}>({
-    assets: [],
-    includedCryptoIds: new Set(),
-    excludedCryptoIds: new Set(),
-});
-
-export interface AssetOptionsContextProps {
-    includedCryptoIds: Set<CryptoId>;
-    excludedCryptoIds: Set<CryptoId>;
-    children: ReactNode;
-}
-
-export function AssetOptionsProvider({
-    includedCryptoIds,
-    excludedCryptoIds,
-    children,
-}: AssetOptionsContextProps) {
-    const { buildAssetOptions } = useTradingAssets();
-    const contextValue = useMemo(() => {
-        const { assets } = buildAssetOptions({ includedCryptoIds });
-
-        return { assets, includedCryptoIds, excludedCryptoIds };
-    }, [buildAssetOptions, excludedCryptoIds, includedCryptoIds]);
-
-    return (
-        <AssetOptionsContext.Provider value={contextValue}>{children}</AssetOptionsContext.Provider>
-    );
-}
-
-export const useAssetsContext = () =>
-    useContext(AssetOptionsContext) ??
-    throwError(`Can't use useAssetsContext outside of AssetOptionsProvider`);
