+++ packages/suite/src/views/wallet/trading/common/TradingForm/TradingFormInput/TradingFormInputBuyAsset/AssetPickerModal/hooks/useBuildTradingAssetOptions.ts
@@ -8,13 +8,14 @@ import {
     filterTradeableAssetsBySearch,
     orderTradeableAssetsByOwnership,
     usePreferredCurrencyUsdThreshold,
+    useTradingAssets,
 } from '@suite-common/trading';
 import { type NetworkSymbol, networkSymbolCollection } from '@suite-common/wallet-config';
 
 import { useSelector } from 'src/hooks/suite';
 import { selectTradeableAssetBalances } from 'src/selectors/wallet/tradeableAssetBalancesSelectors';
 
-import { useAssetsContext } from '../../AssetOptionsContext';
+import { useAssetsContext } from '../../../TradingFormInputAssetPicker';
 
 export type TradingAssetListItem = {
     asset: TradingAssetOption;
@@ -40,14 +41,16 @@ export function useBuildTradingAssetOptions({
     search,
     networkSymbol,
 }: UseBuildTradingAssetOptionsProps) {
-    const { assets, excludedCryptoIds } = useAssetsContext();
+    const { includedCryptoIds, excludedCryptoIds } = useAssetsContext();
+    const { buildAssetOptions } = useTradingAssets();
     const balances = useSelector(selectTradeableAssetBalances);
     const preferredCurrencyUsdThreshold = usePreferredCurrencyUsdThreshold();
 
-    const includedAssets = useMemo(
-        () => assets.filter(asset => !excludedCryptoIds.has(asset.id)),
-        [assets, excludedCryptoIds],
-    );
+    const includedAssets = useMemo(() => {
+        const { assets } = buildAssetOptions({ includedCryptoIds });
+
+        return assets.filter(asset => !excludedCryptoIds.has(asset.id));
+    }, [buildAssetOptions, includedCryptoIds, excludedCryptoIds]);
 
     const searchIndex = useMemo(
         () =>
