+++ packages/suite/src/views/wallet/trading/common/TradingForm/TradingFormInput/TradingFormInputSellAsset/TradingFormInputSellAsset.tsx
@@ -4,8 +4,11 @@ import { type CryptoId } from 'invity-api';
 
 import { useTradingAssetPickerModal } from 'src/hooks/wallet/trading/form/common/useTradingAssetPickerModal';
 
-import { AssetPickerInput, type AssetPickerInputProps } from '../TradingFormInputAssetPicker';
-import { AssetOptionsProvider } from './AssetOptionsContext';
+import {
+    AssetOptionsProvider,
+    AssetPickerInput,
+    type AssetPickerInputProps,
+} from '../TradingFormInputAssetPicker';
 import { AssetPickerModal, type AssetPickerModalProps } from './AssetPickerModal/AssetPickerModal';
 
 export interface TradingFormInputSellAssetProps {
